Fixed bug with internal commands
The bug was there since the very beginning of internal commands, not
related to any recent changes. For some reason the C compiler does
not complain when calling a function pointer instead of a function
without dereferencing it. cd initially was the very first function
in the array, so (*fcall)(args) did work and always call the 1st
function. Then ':' got the 1st function (doing nothing) and it worked
as well. Correct is (*fcall[funcNr])(args). Now the function pointer
is dereferenced properly, fcall[funcNr] is a function pointer,

Looks like I stumbled across an archaic design error in C.